remembering current nav waypoint
Arducopter 3.1.1 (and earlier) versions would re-initialize the current
nav waypoint whenever AUTO mode was entered.  This is detrimental to
switching between STABILIZED and AUTO repeatedly during a flight.

set_mode() to AUTO no longer reinitializes the current waypoint.
Instead that is done in the arm_motors_check when the motors are armed.
This allows the waypoints to be initialized properly before flight, and
allows allows a manual reset by disarming and rearming.